feat: export new features
