+++ src/index.ts
@@ -1,2 +1,3 @@
 export type { CatchExceptionOptions } from './lib';
-export { CatchException, catchException, Logr } from './lib';
+export { CatchException, catchException, Loggable, Logr, Traceable } from './lib';
+export { AsyncTraceStorage } from './lib';
