+++ src/lib/core/decorators/index.ts
@@ -1,1 +1,3 @@
 export { CatchException } from './catch-exception';
+export { Loggable } from './loggable';
+export { Traceable } from './traceable';
