+++ src/lib/core/decorators/loggable/index.ts
@@ -1,0 +1,1 @@
+export * from './loggable';
