+++ src/lib/core/decorators/traceable/index.ts
@@ -1,0 +1,1 @@
+export * from './traceable';
