+++ src/lib/core/types/async-trace.types.ts
@@ -1,0 +1,8 @@
+import { RegisteredErrorDTO } from '@core/dtos';
+
+export type AsyncTrace = {
+  id?: string;
+  correlationId?: string;
+  causationId?: string;
+  registeredError?: RegisteredErrorDTO | undefined;
+};
