+++ src/lib/core/types/index.ts
@@ -1,2 +1,3 @@
+export type { AsyncTrace } from './async-trace.types';
 export type { CatchExceptionOptions } from './catch-exception-options.types';
 export type { CommonOptions } from './common-options.types';
