+++ src/lib/index.ts
@@ -1,6 +1,7 @@
-export { CatchException } from '@core/decorators';
+export { CatchException, Loggable, Traceable } from '@core/decorators';
 export { catchException } from '@core/functions';
 export { Logr } from '@core/services';
-export type { CatchExceptionOptions } from '@core/types';
+export { AsyncTraceStorage } from '@core/storages';
+export type { AsyncTrace, CatchExceptionOptions } from '@core/types';
 export { LoggerStdout } from '@infra/stdout';
 export { LoggerWinston } from '@infra/winston';
